+++ lib/cukedep/constants.rb
@@ -3,7 +3,7 @@
 
 module Cukedep # Module used as a namespace
   # The version number of the gem.
-  Version = '0.0.7'
+  Version = '0.0.8'
 
   # Brief description of the gem.
   Description = 'Manage dependencies between Cucumber feature files'
@@ -11,7 +11,7 @@ module Cukedep # Module used as a namespace
   # Constant Cukedep::RootDir contains the absolute path of Rodent's
   # root directory. Note: it also ends with a slash character.
   unless defined?(RootDir)
-  # The initialisation of constant RootDir is guarded in order 
+  # The initialisation of constant RootDir is guarded in order
   # to avoid multiple initialisation (not allowed for constants)
 
     # The root folder of Cukedep.
@@ -20,7 +20,7 @@ module Cukedep # Module used as a namespace
       rootdir = Pathname(__FILE__).dirname.parent.parent.expand_path
       rootdir.to_s + '/'	# Append trailing slash character to it
     end
-    
+
     # The file name for the user's settings
     YMLFilename = '.cukedep.yml'
   end